requirements.txt
google-analytics-data==0.18.7
stripe==7.12.0
pandas==2.2.0
python-dotenv==1.0.0
openai==1.12.0
streamlit==1.31.0
matplotlib==3.8.2
schedule==1.2.0
loguru==0.7.2
pytest==8.0.0
tweepy==4.14.0